feat: add admin filter, add course with arrays and file image, change enrollments api behavior, add user seeder

+++ app/Config/Routes.php
@@ -18,10 +18,19 @@ $routes->group("api", ["filter" => "cors"], function ($routes) {
         $routes->delete("/", "AuthenticationController::logout");
     });
 
+    $routes->group("courses", function ($routes) {
+        $routes->post("/", "CourseController::store", ["filter" => ["auth", "admin"]]);
+        $routes->get("/", "CourseController::index");
+        // $routes->get("(:any)", "CourseController::image/$1");
+        $routes->put("(:segment)", "CourseController::update/$1", ["filter" => ["auth", "admin"]]);
+        $routes->delete("(:segment)", "CourseController::delete/$1", ["filter" => ["auth", "admin"]]);
+    });
+
     $routes->group("enrollments", ["filter" => "auth"], function ($routes) {
         $routes->post("/", "EnrollmentController::store");
         $routes->get("/", "EnrollmentController::index");
-        $routes->delete("/", "EnrollmentController::destroy");
+        $routes->get("all", "EnrollmentController::all", ["filter" => "admin"]);
+        $routes->delete("(:segment)", "EnrollmentController::destroy/$1", ["filter" => "admin"]);
     });
 
     $routes->group("certificates", function ($routes) {
